+++ internal/server/furnace.go
@@ -146,8 +146,7 @@ func (h *hub) openFurnace(t *tracked, x, y, z int) {
 	if h.nextWin > 100 {
 		h.nextWin = 1
 	}
-	t.winID, t.winPos, t.winKind = h.nextWin, pos, winFurnace
-	h.vib(pos.dim, freqContainerOpen, pos.x, pos.y, pos.z, t.p.eid)
+	t.winID, t.winPos, t.winKind = h.nextWin, pos, winFurnace // no opener counter: no CONTAINER_OPEN
 	if f.viewers == nil {
 		f.viewers = map[int32]*[4]int{}
 	}